Move special methods from spech5 dataset to commonh5 dataset
